requirements.txt
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.3
opencv-python==4.9.0.80
pillow==10.3.0
numpy<2
